+++ libs/render/include/render/bindings.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <exo/collections/span.h>
+#include "exo/collections/span.h"
 
 #include "render/ring_buffer.h"
 #include "render/vulkan/commands.h"
@@ -14,7 +14,7 @@ inline exo::Span<u8> bind_shader_options(
 	auto [p_options, offset] =
 		ring_buffer.allocate(options_len, 0x40); // 0x10 enough on AMD, should probably check device features
 	const auto &descriptor = device.find_or_create_uniform_descriptor(ring_buffer.buffer, options_len);
-	cmd.bind_uniform_set(descriptor, offset);
+	cmd.bind_uniform_set(descriptor, u32(offset));
 	return p_options;
 }
